% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkReactR.R
\name{vtkReactR-shiny}
\alias{vtkReactR-shiny}
\alias{vtkViewOutput}
\alias{renderVtkView}
\alias{vtkReactR_html}
\title{Shiny bindings for vtkReactR}
\usage{
vtkViewOutput(outputId, width = "100\%", height = "400px")

renderVtkView(expr, env = parent.frame(), quoted = FALSE)

vtkReactR_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a vtkReactR}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using vtkReactR within Shiny
applications and interactive Rmd documents.
}
